Move share settings to a new file.
I moved it to a new file but I am not sure what to do with the _CloseButton & _Section. I currently have them both in share_settings.dart and & settings.dart.
