+++ app/lib/routes/settings.dart
@@ -1,13 +1,12 @@
-import 'dart:developer';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:modal_bottom_sheet/modal_bottom_sheet.dart';
-import 'package:mooovy/api.dart';
 import 'package:provider/provider.dart';
 import 'package:mooovy/local_storage.dart';
 import 'package:mooovy/bike/models.dart';
 import 'package:mooovy/bike/bike.dart';
+import 'package:mooovy/widgets/settings/share_settings.dart';
 
 class Settings extends StatelessWidget {
   const Settings({required this.ios, required this.bike, super.key});
@@ -72,109 +71,6 @@ class Settings extends StatelessWidget {
                 ),
               ],
             ),
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-class ShareSettings extends StatelessWidget {
-  const ShareSettings({required this.bike, super.key});
-  final Bike bike;
-
-  @override
-  Widget build(BuildContext context) {
-    return CupertinoScaffold(
-      overlayStyle: SystemUiOverlayStyle.dark,
-      body: Scaffold(
-        appBar: CupertinoNavigationBar(
-          backgroundColor: CupertinoTheme.of(context).scaffoldBackgroundColor,
-          leading: Container(),
-          middle: const Text('Manage sharing'),
-          trailing: _CloseButton(onPressed: () => Navigator.pop(context)),
-        ),
-        body: SafeArea(
-          child: Column(
-            children: [
-              _Section(
-                title: 'Share ${bike.name}',
-                children: [
-                  Column(
-                    crossAxisAlignment: CrossAxisAlignment.start,
-                    children: [
-                      Form(
-                        child: TextFormField(
-                          decoration: const InputDecoration(
-                            labelText: 'Email',
-                          ),
-                          onChanged: (value) {
-                            log(value);
-                          },
-                        ),
-                      ),
-                      ElevatedButton(
-                        onPressed: () {
-                          log('Share bike pressed');
-                        },
-                        child: const Text('Share'),
-                      ),
-                    ],
-                  ),
-                ],
-              ),
-              _Section(
-                title: 'Manage share holders',
-                children: [
-                  Row(
-                    children: [
-                      const Expanded(
-                        child: Text('Share holder 1'),
-                      ),
-                      const Text("1 day"),
-                      IconButton(
-                        icon: const Icon(Icons.close),
-                        onPressed: () async {
-                          var api = obtainApiClient();
-                          var list = await api?.getCurrentShares(bike.id);
-                          log(list[0]["email"].toString());
-
-                          log('Removed share holder 1');
-                        },
-                      ),
-                    ],
-                  ),
-                  Row(
-                    children: [
-                      const Expanded(
-                        child: Text('Share holder 2'),
-                      ),
-                      const Text("Forever"),
-                      IconButton(
-                        icon: const Icon(Icons.close),
-                        onPressed: () {
-                          log('Removed share holder 2');
-                        },
-                      ),
-                    ],
-                  ),
-                  Row(
-                    children: [
-                      const Expanded(
-                        child: Text('Share holder 3'),
-                      ),
-                      const Text("1 week"),
-                      IconButton(
-                        icon: const Icon(Icons.close),
-                        onPressed: () {
-                          log('Removed share holder 3');
-                        },
-                      ),
-                    ],
-                  ),
-                ],
-              ),
-            ],
           ),
         ),
       ),
